Fix wallet drawer close tracking
Move useUmami into WalletDrawer component rendered only when open,
tracking at Drawer onClose boundary so the event fires for all
close paths (X button, click-outside, Escape). Also prevents
useSearchParams from being called during static prerendering.
